[groovy-splash] Improved segments & layout system
This is a big commit, in part because many of the improvements to the
segments were made in order to support the new layout system.

Segments are now defined as functions which take an optional HEIGHT
argument. The WIDTH argument is removed, and the NO-DRAW argument is
folded into the HEIGHT by passing nil. Segments also now return a
range (MIN . MAX), where MAX may be nil, representing an
arbitrarily-tall segment.

The layout system is now able to accommodate a wider range of frame
heights, and works as follows:

  - If there is enough space for every segment with a maximum height to
    have its maximum height, then the extra space is shared equally
    among segments with no maximum height.

  - If there is enough space for every segment to have its minimum
    height, each segment is given its maximum height if it has one and
    its minimum otherwise. Then the height of the first segment is
    reduced until the remaining segments fit into the screen. If
    reducing the first segment to its minimum height does not cause the
    remaining segments to fit, the process is repeated with the second
    segment, continuing until all segments fit on the screen.

  - If there is not enough space for every segment, just give all of them
    their minimum height.

It is assumed that the user will adjust their layout to suit their
initial frame size (or vice-versa); the fallback layout strategies are
strictly best-effort. We also note that less-useful segments, such as
banners, tend to be near the top, and more-useful segments, such as the
recentf list, tend to be towards the bottom.

The groovy-fill segment can now be customized with different stripe
characters and different colors.

The rule segment now displays correctly.

The logo segment is now able to adjust its height over a range, and its
maximum height can be customized.

The oracle segment now supports customization of the bracketing
characters.

The recentf segment now supports showing as many files as it can. If
there are not enough files to fill the space, it will function as if it
was followed by fill-blank.